Fix: 감각별 분포도에서 발생한 이슈 원인 파악 완료

+++ app/src/main/java/com/mangpo/taste/view/model/PercentageOfCategory.kt
@@ -4,7 +4,7 @@ import android.graphics.drawable.Drawable
 
 data class PercentageOfCategory (
     val sense: String,
-    val percentage: Int,
+    val percentage: Float,
     val color: Int,
     val character: Drawable,
     val graph: Drawable,
