+++ data/src/main/java/com/mangpo/data/model/getStat/PercentageOfCategoryDTO.kt
@@ -1,10 +1,10 @@
 package com.mangpo.data.model.getStat
 
 data class PercentageOfCategoryDTO(
-    val AMBIGUOUS: Int,
-    val HEARING: Int,
-    val SIGHT: Int,
-    val SMELL: Int,
-    val TASTE: Int,
-    val TOUCH: Int
+    val AMBIGUOUS: Float,
+    val HEARING: Float,
+    val SIGHT: Float,
+    val SMELL: Float,
+    val TASTE: Float,
+    val TOUCH: Float
 )
