+++ domain/src/main/java/com/mangpo/domain/model/getStat/PercentageOfCategoryEntity.kt
@@ -1,10 +1,10 @@
 package com.mangpo.domain.model.getStat
 
 data class PercentageOfCategoryEntity(
-    val AMBIGUOUS: Int,
-    val HEARING: Int,
-    val SIGHT: Int,
-    val SMELL: Int,
-    val TASTE: Int,
-    val TOUCH: Int
+    val AMBIGUOUS: Float,
+    val HEARING: Float,
+    val SIGHT: Float,
+    val SMELL: Float,
+    val TASTE: Float,
+    val TOUCH: Float
 )
